normalizes inventory property names, adds enums file
