+++ app/src/util/enums.ts
@@ -1,0 +1,9 @@
+export enum InventoryType {
+  GPC_BASIC = "gpc_basic",
+  GPC_BASIC_PLUS = "gpc_basic_plus",
+}
+
+export enum GlobalWarmingPotentialType {
+  ar6 = "ar6",
+  ar5 = "ar5",
+}
